feat: Implement SettingsActivity with API key management and transcription model selection
- Added SettingsActivity for managing OpenAI API key and transcription settings.
- Integrated SharedPrefsUtils for secure storage of API keys and user preferences.
- Implemented UI components for inputting and testing API keys, selecting transcription models, and configuring prompts.
- Added functionality to reset prompts and language settings to default values.
- Created utility class FileProcessingManager for handling audio file processing.
- Established a theme with custom colors and typography for the application.
- Included example unit test for basic functionality verification.
