+++ app/src/main/java/app/hdev/io/aitranscribe/AITranscriptionApp.kt
@@ -1,0 +1,7 @@
+package app.hdev.io.aitranscribe
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class AITranscriptionApp : Application()
