--- app/src/main/java/com/example/aiaudiotranscription/AITranscriptionApp.kt
@@ -1,7 +1,0 @@
-package com.example.aiaudiotranscription
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class AITranscriptionApp : Application()
