--- app/src/main/java/com/example/aiaudiotranscription/di/AppModule.kt
@@ -1,35 +1,0 @@
-package com.example.aiaudiotranscription.di
-
-import android.content.Context
-import com.example.aiaudiotranscription.api.OpenAiApiService
-import com.example.aiaudiotranscription.api.RetrofitClient
-import com.example.aiaudiotranscription.data.TranscriptionDbHelper
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-
-    @Provides
-    @Singleton
-    fun provideContext(@ApplicationContext context: Context): Context {
-        return context
-    }
-
-    @Provides
-    @Singleton
-    fun provideTranscriptionDbHelper(@ApplicationContext context: Context): TranscriptionDbHelper {
-        return TranscriptionDbHelper(context)
-    }
-
-    @Provides
-    @Singleton
-    fun provideWhisperApiService(@ApplicationContext context: Context): OpenAiApiService {
-        return RetrofitClient.create(context).create(OpenAiApiService::class.java)
-    }
-}
