+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">AI Transcription</string>
+    <string name="app_name">AI Universal Transcriber</string>
 </resources>
